Tests: Add admin e2e drag-reorder coverage to nested-list scenario
- Add `order` int property and `ordered` scope to Item
- Declare `scope: 'ordered'` on Widget.items so eager loads sort
- Schema sets `draggable: true` and `orderKey: 'order'` on items
- New drag-reorder.ts asserts DOM swap, save, reload, DB order
- Split index.spec.ts into barrel + crud.ts, mirrors assets layout
