+++ tests/e2e/nested-list/models/Item.ts
@@ -1,10 +1,12 @@
 import type { ModelProperties } from '@ditojs/server'
 import { Model } from '@ditojs/server'
+import type { QueryBuilder } from 'objection'
 
 export interface Item {
   id: number
   widgetId: number
   label: string
+  order: number | null
 }
 
 export class Item extends Model {
@@ -13,6 +15,12 @@ export class Item extends Model {
     // it on inserted children from the parent relation — the request body
     // doesn't carry it.
     widgetId: { type: 'integer', index: true, nullable: true },
-    label: { type: 'string', required: true }
+    label: { type: 'string', required: true },
+    order: { type: 'integer', nullable: true }
+  }
+
+  static override scopes = {
+    ordered: (query: QueryBuilder<Item>) =>
+      query.orderBy('order').orderBy('id')
   }
 }
